+++ lib/rpc/server/http.js
@@ -1,15 +1,13 @@
-const Rpc = require('./ethRPC.js');
 const Express = require('express');
 const bodyParser = require('body-parser');
 const async = require('async');
 
-module.exports = XhrRpc;
+module.exports = HttpRpc;
 
-function XhrRpc(opts){
 
-  var app = this.app = opts.app;
+function HttpRpc(rpc){
+
   var server = this.server = Express();
-  var rpc = this.rpc = opts.app.rpc;
 
   // parse application/json even if it doesnt say its json (*/* does not work)
   server.use(bodyParser.json({ type: 'application/*' }));
@@ -90,9 +88,9 @@ function XhrRpc(opts){
 }
 
 //the function that starts the application
-XhrRpc.prototype.start = function(options, done) {
+HttpRpc.prototype.start = function(options, done) {
   var port = options.port || 8080;
-  console.log('XhrRpcServer - opening: /:' + port);
+  console.log('HttpRpcServer - opening: /:' + port);
   this.httpServer = this.server.listen(port);
   process.once('SIGTERM', this.stop.bind(this));
   process.once('SIGINT', this.stop.bind(this));
@@ -100,7 +98,7 @@ XhrRpc.prototype.start = function(options, done) {
   done();
 };
 
-XhrRpc.prototype.stop = function(done) {
+HttpRpc.prototype.stop = function(done) {
   this.httpServer.close();
   if (done) done();
 };
